[Bug] Keep the mock-vllm app under the structure gate's file limit
Adding /classify to tools/mock-vllm/app.py took the file from 796 to
821 lines, past the 800-line limit the agent structure check enforces,
so Core quality / Full pre-commit checks went red on the E2E commit.

Move the http_classify stand-in into classify.py and mount it through an
APIRouter the way provider_boundary already is. The endpoint, its window
size and its scores are unchanged; app.py only imports and includes the
router. The Dockerfile copies the new module so the E2E image still
serves /classify, and the README lists the endpoint and the module.
