+++ tools/mock-vllm/Dockerfile
@@ -9,7 +9,7 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
 COPY requirements.txt ./
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY app.py chat_request.py provider_boundary.py provider_contract.py ./
+COPY app.py chat_request.py classify.py provider_boundary.py provider_contract.py ./
 COPY schema_contract.json ./
 
 EXPOSE 8000
